fix(watcher/bench): use %ld not %lld (ESP-IDF newlib-nano doesn't support %lld)
Previous patch (b69e1d6) used %lld + (long long) casts for the timing
us values. On ESP32-S3 with the default ESP-IDF newlib-nano stdio, the
'll' length modifier is silently dropped, leaving the literal 'ld' in
output and corrupting all subsequent printf args. Symptom: bench output
showed 'at_face=ld invoke_to_result=ld ...' with garbage 'got=' and
'score=' values.

us timings fit comfortably in long (32-bit signed → ~35 minutes), so
downgrading the casts is fine.

+++ main/boards/sensecap-watcher/face_serial_handler.cc
@@ -361,10 +361,10 @@ int FaceSerialHandler::CmdBench(int argc, char** argv) {
     if (!ok) {
         // Still print the timing — useful to see how long we waited before timeout.
         printf("{\"ok\":false,\"error\":\"no_face_or_timeout\","
-               "\"timing_us\":{\"at_face\":%lld,\"invoke_to_result\":%lld,"
-               "\"teardown\":%lld,\"total\":%lld}}\n",
-               (long long)t.at_face_us, (long long)t.invoke_to_result_us,
-               (long long)t.teardown_us, (long long)t.total_us);
+               "\"timing_us\":{\"at_face\":%ld,\"invoke_to_result\":%ld,"
+               "\"teardown\":%ld,\"total\":%ld}}\n",
+               (long)t.at_face_us, (long)t.invoke_to_result_us,
+               (long)t.teardown_us, (long)t.total_us);
         return 0;
     }
 
@@ -383,13 +383,13 @@ int FaceSerialHandler::CmdBench(int argc, char** argv) {
 
     printf("{\"ok\":true,\"format\":\"float32_le_b64\",\"dim\":%d,"
            "\"score\":%d,\"quality\":%.3f,"
-           "\"timing_us\":{\"at_face\":%lld,\"invoke_to_result\":%lld,"
-           "\"teardown\":%lld,\"total\":%lld},"
+           "\"timing_us\":{\"at_face\":%ld,\"invoke_to_result\":%ld,"
+           "\"teardown\":%ld,\"total\":%ld},"
            "\"embedding\":\"%.*s\"}\n",
            FACE_EMBEDDING_DIM,
            t.face_score, t.face_quality,
-           (long long)t.at_face_us, (long long)t.invoke_to_result_us,
-           (long long)t.teardown_us, (long long)t.total_us,
+           (long)t.at_face_us, (long)t.invoke_to_result_us,
+           (long)t.teardown_us, (long)t.total_us,
            (int)b64_len, (const char*)b64);
     return 0;
 }
